requirements.txt
flask-restplus==0.9.2
Flask-SQLAlchemy==2.1
Werkzeug==0.16.1
markupsafe==2.0.1
defaultdict
uuid
json
datetime
